fix UTF8 truncation regression

+++ test/iteration/iteration.c
@@ -59,9 +59,9 @@ static void   test( mulle_utf32_t text[ 4])
    struct buffer   buffer8;
    mulle_utf32_t   *utf32;
    mulle_utf16_t   *utf16;
-   char    *utf8;
+   char            *utf8;
    mulle_utf32_t   c;
-   size_t    i;
+   unsigned int    i;
 
    memset( &buffer16, 0, sizeof( buffer16));
    memset( &buffer8, 0, sizeof( buffer8));
@@ -127,7 +127,7 @@ static void   test( mulle_utf32_t text[ 4])
 static void   stress_test()
 {
    mulle_utf32_t   text[ 4];
-   size_t    i;
+   unsigned int    i;
    mulle_utf32_t   mask;
 
    mask = 0;
